# src/main/resources/openapi/jobly_workflow_openapi.yaml
openapi: 3.0.4
info:
  title: Jobly API
  description: |-
    This service will be used to manage the Jobly application.
  version: 1.0.0
servers:
  - url: https://jobly.com/api/v1
paths:
...
  /job-offers:
    get:
      summary: Get all job offers
      tags:
        - Job Offers
      operationId: getAllJobOffers
      responses:
        '200':
          description: A list of all job offers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllJobOffersResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedMessage'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenMessage'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorMessage'
  /job-offers/{id}:
    get:
      summary: Get job offer by ID
      tags:
        - Job Offers
      operationId: getJobOfferById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the job offer to retrieve
      responses:
        '200':
          description: Job offer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobOfferDetailsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedMessage'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenMessage'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundMessage'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorMessage'
    put:
      summary: Update an existing job offer
      tags:
        - Job Offers
      operationId: updateJobOffer
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the job offer to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateJobOfferRequest'
      responses:
        '200':
          description: Job offer updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobOffer'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestMessage'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedMessage'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenMessage'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundMessage'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorMessage'
    delete:
      summary: Delete job offer by ID
      tags:
        - Job Offers
      operationId: deleteJobOffer
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the job offer to delete
      responses:
        '204':
          description: Job offer deleted successfully
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestMessage'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedMessage'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenMessage'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundMessage'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorMessage'
  /job-offers/mine:
    get:
      summary: Get job offers owned by the authenticated user
      tags:
        - Job Offers
      operationId: getOwnedJobOffers
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: A list of job offers owned by the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMineJobOffersResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedMessage'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenMessage'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorMessage'
  /job-offers/mine/{id}:
    get:
      summary: Get job offer and its applications owned by the authenticated user
      tags:
        - Job Offers
      operationId: getOwnedJobOfferDetails
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the job offer and its applications to retrieve
      responses:
        '200':
          description: Job offer and applications related to it
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobOfferWithApplicationsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedMessage'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenMessage'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundMessage'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorMessage'
  /job-offers/create:
    post:
      summary: Create a new job offer
      tags:
        - Job Offers
      operationId: createJobOffer
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobOfferRequest'
      responses:
        '201':
          description: Job offer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobOffer'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestMessage'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedMessage'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenMessage'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorMessage'
  /job-offers/{jobOfferId}/apply:
    post:
      summary: Apply to a job offer
      tags:
        - Application
      operationId: applyToJobOffer
      security:
        - bearerAuth: [ ]
      parameters:
        - name: jobOfferId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the job offer to apply to
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationCreateRequest'
      responses:
        '201':
          description: Application submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestMessage'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedMessage'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenMessage'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundMessage'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictMessage'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorMessage'
  /applications/mine:
    get:
      summary: Get applications submitted by the authenticated user
      tags:
        - Application
      operationId: getMyApplications
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: A list of applications submitted by the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMyApplicationsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedMessage'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenMessage'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorMessage'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    DateTime:
      type: string
      format: date-time
      example: '2023-10-01T12:00:00Z'
    SystemGeneratedId:
      type: integer
      format: int64
      example: 1
    BadRequestMessage:
      type: object
      properties:
        status:
          type: integer
          example: 400
        timestamp:
          $ref: '#/components/schemas/DateTime'
        message:
          type: string
          example: Invalid input data
    UnauthorizedMessage:
      type: object
      properties:
        status:
          type: integer
          example: 401
        timestamp:
          $ref: '#/components/schemas/DateTime'
        message:
          type: string
          example: User is unauthorized to access this resource
    ForbiddenMessage:
      type: object
      properties:
        status:
          type: integer
          example: 403
        timestamp:
          $ref: '#/components/schemas/DateTime'
        message:
          type: string
          example: User is forbidden from accessing this resource
    NotFoundMessage:
      type: object
      properties:
        status:
          type: integer
          example: 404
        timestamp:
          $ref: '#/components/schemas/DateTime'
        message:
          type: string
          example: Resource not found
    ConflictMessage:
      type: object
      properties:
        status:
          type: integer
          example: 409
        timestamp:
          $ref: '#/components/schemas/DateTime'
        message:
          type: string
          example: Resource already exists
    InternalServerErrorMessage:
      type: object
      properties:
        status:
          type: integer
          example: 500
        timestamp:
          $ref: '#/components/schemas/DateTime'
        message:
          type: string
          example: An unexpected error occurred on the server
...
    GetAllJobOffersResponse:
      type: object
      properties:
        total:
          type: integer
          example: 5
        jobOffers:
          type: array
          items:
            $ref: '#/components/schemas/JobOfferListObject'
    GetMineJobOffersResponse:
      type: object
      properties:
        total:
          type: integer
          example: 5
        jobOffers:
          type: array
          items:
            $ref: '#/components/schemas/JobOfferListObject'
    JobOfferListObject:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/SystemGeneratedId'
        title:
          type: string
          example: Senior Java Developer
          maxLength: 32
        companyName:
          type: string
          example: Tech Solutions Inc.
          maxLength: 60
        salary:
          type: number
          format: float
          example: 85000.00
        workType:
          $ref: '#/components/schemas/WorkType'
        location:
          type: string
          example: New York, NY
          maxLength: 100
        category:
          $ref: '#/components/schemas/JobOfferCategory'
    JobOffer:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/SystemGeneratedId'
        title:
          type: string
          example: Senior Java Developer
          maxLength: 32
        description:
          type: string
          example: We are looking for a Senior Java Developer with 5+ years of experience...
          maxLength: 1024
        companyName:
          type: string
          example: Tech Solutions Inc.
          maxLength: 60
        salary:
          type: number
          format: float
          example: 85000.00
        yearsOfExperience:
          type: integer
          format: int32
          example: 5
        workType:
          $ref: '#/components/schemas/WorkType'
        location:
          type: string
          example: New York, NY
          maxLength: 100
        contactEmail:
          type: string
          format: email
        contactPhone:
          type: string
          example: +37061234567
          maxLength: 15
        offerStatus:
          $ref: '#/components/schemas/JobOfferStatus'
        createdAt:
          $ref: '#/components/schemas/DateTime'
        updatedAt:
          $ref: '#/components/schemas/DateTime'
        category:
          $ref: '#/components/schemas/JobOfferCategory'
        creator:
          $ref: '#/components/schemas/Creator'
    WorkType:
      type: string
      enum: [ ON_SITE, REMOTE, HYBRID ]
    JobOfferStatus:
      type: string
      enum: [ OPEN, CLOSED ]
    JobOfferCategory:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/SystemGeneratedId'
        name:
          type: string
          example: Software Development
          maxLength: 32
    Creator:
      type: object
      properties:
        firstName:
          type: string
          example: John
          maxLength: 20
        lastName:
          type: string
          example: Doe
          maxLength: 20
    Application:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/SystemGeneratedId'
        comment:
          type: string
          maxLength: 1000
        createdAt:
          $ref: '#/components/schemas/DateTime'
        updatedAt:
          $ref: '#/components/schemas/DateTime'
        applicant:
          $ref: '#/components/schemas/Applicant'
        applicationStatus:
          $ref: '#/components/schemas/ApplicationStatus'
        cvId:
          $ref: '#/components/schemas/SystemGeneratedId'
    ApplicationStatus:
      type: string
      enum: [ PENDING, ACCEPTED, REJECTED, WITHDRAWN ]
    Applicant:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/SystemGeneratedId'
        firstName:
          type: string
          example: Jane
          maxLength: 20
        lastName:
          type: string
          example: Smith
          maxLength: 20
        email:
          type: string
          format: email
    JobOfferWithApplicationsResponse:
      type: object
      properties:
        jobOffer:
          $ref: '#/components/schemas/JobOffer'
        applications:
          type: array
          items:
            $ref: '#/components/schemas/Application'
    JobOfferDetailsResponse:
      type: object
      properties:
        jobOffer:
          $ref: '#/components/schemas/JobOffer'
    CreateJobOfferRequest:
      type: object
      required:
        - title
        - description
        - companyName
        - salary
        - yearsOfExperience
        - workType
        - contactEmail
        - categoryId
      properties:
        title:
          type: string
          example: Senior Java Developer
          maxLength: 32
        description:
          type: string
          example: We are looking for a Senior Java Developer with 5+ years of experience...
          maxLength: 1024
        companyName:
          type: string
          example: Tech Solutions Inc.
          maxLength: 60
        salary:
          type: number
          format: float
          example: 85000.00
        yearsOfExperience:
          type: integer
          format: int32
          example: 5
        workType:
          $ref: '#/components/schemas/WorkType'
        location:
          type: string
          example: New York, NY
          maxLength: 100
        contactEmail:
          type: string
          format: email
        contactPhone:
          type: string
          example: +37061234567
          maxLength: 15
        categoryId:
          $ref: '#/components/schemas/SystemGeneratedId'
    UpdateJobOfferRequest:
      type: object
      required:
        - title
        - description
        - companyName
        - salary
        - yearsOfExperience
        - workType
        - contactEmail
        - categoryId
      properties:
        title:
          type: string
          example: Senior Java Developer
          maxLength: 32
        description:
          type: string
          example: We are looking for a Senior Java Developer with 5+ years of experience...
          maxLength: 1024
        companyName:
          type: string
          example: Tech Solutions Inc.
          maxLength: 60
        salary:
          type: number
          format: float
          example: 85000.00
        yearsOfExperience:
          type: integer
          format: int32
          example: 5
        workType:
          $ref: '#/components/schemas/WorkType'
        location:
          type: string
          example: New York, NY
          maxLength: 100
        contactEmail:
          type: string
          format: email
        contactPhone:
          type: string
          example: +37061234567
          maxLength: 15
        categoryId:
          $ref: '#/components/schemas/SystemGeneratedId'
...
    ApplicationCreateRequest:
      type: object
      properties:
        comment:
          type: string
          example: I am very interested in this position and believe my skills are a great match.
          maxLength: 1000
    GetMyApplicationsResponse:
      type: object
      properties:
        total:
          type: integer
          example: 3
        applications:
          type: array
          items:
            $ref: '#/components/schemas/MyApplicationListObject'
    MyApplicationListObject:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/SystemGeneratedId'
        updatedAt:
          $ref: '#/components/schemas/DateTime'
        jobOffer:
          $ref: '#/components/schemas/JobOfferListObject'
        applicationStatus:
          $ref: '#/components/schemas/ApplicationStatus'
...